Fix a bug of files with missing storage_name
The storage name is generated when created the file. It calls _assure_beam_dir
which actually creates the required directory. This operation can fail, which
causes the file to be registered without storage_name. When a retry is made, the
path is not re-created because the file exists in the database